leave only npm publish logic in "publish.js" and move the rest in "version.js"
