+++ src/index.js
@@ -1,4 +1,6 @@
 import minimist from 'minimist';
+import { every } from 'lodash';
+import version from './npm/version';
 import publish from './npm/publish';
 import release from './github/release';
 import changelog from './github/changelog';
@@ -8,19 +10,28 @@ import config from './config';
 
 const _argv = minimist(process.argv.slice(2));
 
-const defaultArgv = { 'npm-publish': true, 'gh-release': true, changelog: true, 'gh-release-all': false };
+const defaultArgv = {
+  'npm-publish': true,
+  'no-npm-publish': false,
+  'npm-version': true,
+  'gh-release': true,
+  'gh-release-all': false,
+  changelog: true
+};
 
-const argv = (_argv['npm-publish'] || _argv['gh-release'] || _argv['gh-release-all'] || _argv.changelog) ?
-  _argv :
-  defaultArgv;
+const runDefault = every(Object.keys(defaultArgv), arg => typeof _argv[arg] === 'undefined');
+
+const argv = runDefault ? defaultArgv : _argv;
 
 const main = async () => {
   try {
     !config.github.token && await askForToken();
 
-    argv['npm-publish'] && await publish();
+    argv['npm-version'] && await version();
     argv.changelog && await changelog();
     argv['gh-release'] && await release({ all: false });
+    ((runDefault && !argv['no-npm-publish']) || argv['npm-publish']) && await publish();
+
     argv['gh-release-all'] && await release({ all: true });
   } catch (e) {
     onError(e);
